% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sailing_function.R
\name{sailing_function}
\alias{sailing_function}
\title{A Sailing Function}
\usage{
sailing_function(love = TRUE)
}
\arguments{
\item{love}{Do you love sailing? Defaults to TRUE.}
}
\description{
This function allows you to express your love of sailing.
}
\examples{
sailing_function()
}
\keyword{sailing}
